Chapter7/02-ChatGPT-Customer-Service/requirements.txt
line-bot-sdk==2.4.1
Flask==2.2.3
python-dotenv==0.21.1
openai==0.27.0
numpy==1.24.2
pandas==1.5.3